backend/requirements.txt
# iPaper Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# LLM
openai==1.12.0

# Data Validation
pydantic==2.6.0
pydantic-settings==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-multipart==0.0.6
arxiv==2.1.0

# PDF Processing
PyMuPDF==1.25.4

